+++ tests/testthat/test-scale-date.R
@@ -69,9 +69,24 @@ test_that("datetime colour scales work", {
   expect_equal(range(get_layer_data(p)$colour), c("#132B43", "#56B1F7"))
 })
 
-test_that("date(time) scales throw warnings when input is numeric", {
+test_that("date(time) scales throw warnings when input is incorrect", {
   p <- ggplot(data.frame(x = 1, y = 1), aes(x, y)) + geom_point()
 
   expect_snapshot_warning(ggplot_build(p + scale_x_date()))
   expect_snapshot_warning(ggplot_build(p + scale_x_datetime()))
+
+  expect_snapshot(
+    ggplot_build(p + scale_x_date(date_breaks = c(11, 12))),
+    error = TRUE
+  )
+
+  expect_snapshot(
+    ggplot_build(p + scale_x_date(date_minor_breaks = c(11, 12))),
+    error = TRUE
+  )
+
+  expect_snapshot(
+    ggplot_build(p + scale_x_date(date_labels = c(11, 12))),
+    error = TRUE
+  )
 })